+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Version 1.8 (not yet released)
 -----------
 
+* Added Latitude and Longitude fields to address actions.
+* New action: 'Add attachment to bulk mail'.
 
 Version 1.7
 -----------